requirements.txt
asgiref==3.7.2
boto3==1.28.15
botocore==1.31.15
cffi==1.15.1
cryptography==41.0.3
Django==4.2.3
django-allow-cidr==0.7.1
django-storages==1.13.2
gunicorn==21.2.0
jmespath==1.0.1
packaging==23.1
Pillow==10.0.0
psycopg2==2.9.6
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
urllib3==1.26.16
